refactor: changed services directory architecture
