--- src/services/getProducts.js
@@ -1,11 +1,0 @@
-import axios from "axios";
-export const getProducts = async () => {
-  try {
-    const {
-      data: { products },
-    } = await axios.get("/api/products");
-    return products;
-  } catch (error) {
-    console.error(error);
-  }
-};
