--- src/services/postNewUser.js
@@ -1,11 +1,0 @@
-import axios from "axios";
-
-export const postNewUser = async (firstName, lastName, email, password) => {
-  try {
-    const {data,status} = await axios.post("/api/auth/signup",{firstName,lastName,email,password});
-    return {data,status};
-  } catch (error) {
-    console.error("failed to save user to db",error);
-    return error;
-  }
-}
